some more refactor
