+++ crates/node/src/comms/broadcast/state_delta/events.rs
@@ -1,0 +1,70 @@
+use calimero_context_primitives::client::ContextClient;
+use calimero_node_primitives::client::NodeClient;
+use calimero_primitives::context::ContextId;
+use calimero_primitives::events::{
+    ContextEvent, ContextEventPayload, ExecutionEvent, NodeEvent, StateMutationPayload,
+};
+use calimero_primitives::hash::Hash;
+use calimero_primitives::identity::PublicKey;
+use eyre::Result;
+use tracing::{debug, warn};
+
+pub(super) async fn execute_event_handlers_parsed(
+    context_client: &ContextClient,
+    context_id: &ContextId,
+    our_identity: &PublicKey,
+    events_payload: &[ExecutionEvent],
+) -> Result<()> {
+    for event in events_payload {
+        if let Some(handler_name) = &event.handler {
+            debug!(
+                %context_id,
+                event_kind = %event.kind,
+                handler_name = %handler_name,
+                "Executing handler for event"
+            );
+
+            match context_client
+                .execute(
+                    context_id,
+                    our_identity,
+                    handler_name.clone(),
+                    event.data.clone(),
+                    vec![],
+                    None,
+                )
+                .await
+            {
+                Ok(_handler_response) => {
+                    debug!(handler_name = %handler_name, "Handler executed successfully");
+                }
+                Err(err) => {
+                    warn!(handler_name = %handler_name, error = %err, "Handler execution failed");
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
+
+pub(super) fn emit_state_mutation_event_parsed(
+    node_client: &NodeClient,
+    context_id: &ContextId,
+    root_hash: Hash,
+    events_payload: Vec<ExecutionEvent>,
+) -> Result<()> {
+    let state_mutation = ContextEvent {
+        context_id: *context_id,
+        payload: ContextEventPayload::StateMutation(StateMutationPayload::with_root_and_events(
+            root_hash,
+            events_payload,
+        )),
+    };
+
+    if let Err(e) = node_client.send_event(NodeEvent::Context(state_mutation)) {
+        warn!(%context_id, error = %e, "Failed to emit state mutation event to WebSocket clients");
+    }
+
+    Ok(())
+}
