+++ crates/node/src/comms/broadcast/state_delta/mod.rs
@@ -2,23 +2,26 @@
 //!
 //! **SRP**: This module has ONE job - process state deltas from peers using DAG
 
-use calimero_context_primitives::client::ContextClient;
 use calimero_crypto::Nonce;
 use calimero_network_primitives::client::NetworkClient;
-use calimero_node_primitives::client::NodeClient;
 use calimero_primitives::context::ContextId;
-use calimero_primitives::events::{
-    ContextEvent, ContextEventPayload, ExecutionEvent, NodeEvent, StateMutationPayload,
-};
+use calimero_primitives::events::ExecutionEvent;
 use calimero_primitives::hash::Hash;
 use calimero_primitives::identity::PublicKey;
-use calimero_storage::action::Action;
 use eyre::{bail, OptionExt, Result};
 use libp2p::PeerId;
-use tracing::{debug, info, warn};
+use tracing::{info, warn};
 
 use crate::delta_store::DeltaStore;
 use crate::utils::choose_stream;
+
+mod events;
+mod key_share;
+mod missing;
+
+use events::{emit_state_mutation_event_parsed, execute_event_handlers_parsed};
+use key_share::request_key_share_with_peer;
+use missing::request_missing_deltas;
 
 /// Handles state delta received from a peer (DAG-based)
 ///
@@ -412,314 +415,4 @@ pub async fn handle_state_delta(
     }
 
     Ok(())
-}
-
-async fn execute_event_handlers_parsed(
-    context_client: &ContextClient,
-    context_id: &ContextId,
-    our_identity: &PublicKey,
-    events_payload: &[ExecutionEvent],
-) -> Result<()> {
-    for event in events_payload {
-        if let Some(handler_name) = &event.handler {
-            debug!(
-                %context_id,
-                event_kind = %event.kind,
-                handler_name = %handler_name,
-                "Executing handler for event"
-            );
-
-            match context_client
-                .execute(
-                    context_id,
-                    our_identity,
-                    handler_name.clone(),
-                    event.data.clone(),
-                    vec![],
-                    None,
-                )
-                .await
-            {
-                Ok(_handler_response) => {
-                    debug!(
-                        handler_name = %handler_name,
-                        "Handler executed successfully"
-                    );
-                }
-                Err(err) => {
-                    warn!(
-                        handler_name = %handler_name,
-                        error = %err,
-                        "Handler execution failed"
-                    );
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn emit_state_mutation_event_parsed(
-    node_client: &NodeClient,
-    context_id: &ContextId,
-    root_hash: Hash,
-    events_payload: Vec<ExecutionEvent>,
-) -> Result<()> {
-    let state_mutation = ContextEvent {
-        context_id: *context_id,
-        payload: ContextEventPayload::StateMutation(StateMutationPayload::with_root_and_events(
-            root_hash,
-            events_payload,
-        )),
-    };
-
-    if let Err(e) = node_client.send_event(NodeEvent::Context(state_mutation)) {
-        warn!(
-            %context_id,
-            error = %e,
-            "Failed to emit state mutation event to WebSocket clients"
-        );
-    }
-
-    Ok(())
-}
-
-async fn request_missing_deltas(
-    network_client: NetworkClient,
-    sync_timeout: std::time::Duration,
-    context_id: ContextId,
-    missing_ids: Vec<[u8; 32]>,
-    source: PeerId,
-    our_identity: PublicKey,
-    delta_store: DeltaStore,
-) -> Result<()> {
-    use calimero_node_primitives::sync::{InitPayload, MessagePayload, StreamMessage};
-
-    let mut stream = network_client.open_stream(source).await?;
-
-    let mut to_fetch = missing_ids;
-    let mut fetched_deltas: Vec<(calimero_dag::CausalDelta<Vec<Action>>, [u8; 32])> = Vec::new();
-    let mut fetch_count = 0;
-
-    while !to_fetch.is_empty() {
-        let current_batch = to_fetch.clone();
-        to_fetch.clear();
-
-        for missing_id in current_batch {
-            fetch_count += 1;
-
-            info!(
-                %context_id,
-                delta_id = ?missing_id,
-                total_fetched = fetch_count,
-                "Requesting missing parent delta from peer"
-            );
-
-            let request_msg = StreamMessage::Init {
-                context_id,
-                party_id: our_identity,
-                payload: InitPayload::DeltaRequest {
-                    context_id,
-                    delta_id: missing_id,
-                },
-                next_nonce: {
-                    use rand::Rng;
-                    rand::thread_rng().gen()
-                },
-            };
-
-            crate::sync::stream::send(&mut stream, &request_msg, None).await?;
-
-            let timeout_budget = sync_timeout / 3;
-            match crate::sync::stream::recv(&mut stream, None, timeout_budget).await? {
-                Some(StreamMessage::Message {
-                    payload: MessagePayload::DeltaResponse { delta },
-                    ..
-                }) => {
-                    let storage_delta: calimero_storage::delta::CausalDelta =
-                        borsh::from_slice(&delta)?;
-
-                    info!(
-                        %context_id,
-                        delta_id = ?missing_id,
-                        action_count = storage_delta.actions.len(),
-                        "Received missing parent delta"
-                    );
-
-                    let dag_delta = calimero_dag::CausalDelta {
-                        id: storage_delta.id,
-                        parents: storage_delta.parents.clone(),
-                        payload: storage_delta.actions,
-                        hlc: storage_delta.hlc,
-                        expected_root_hash: storage_delta.expected_root_hash,
-                    };
-
-                    fetched_deltas.push((dag_delta, missing_id));
-
-                    for parent_id in &storage_delta.parents {
-                        if *parent_id == [0; 32] {
-                            continue;
-                        }
-                        if !delta_store.has_delta(parent_id).await
-                            && !to_fetch.contains(parent_id)
-                            && !fetched_deltas.iter().any(|(d, _)| d.id == *parent_id)
-                        {
-                            to_fetch.push(*parent_id);
-                        }
-                    }
-                }
-                Some(StreamMessage::Message {
-                    payload: MessagePayload::DeltaNotFound,
-                    ..
-                }) => {
-                    warn!(%context_id, delta_id = ?missing_id, "Peer doesn't have requested delta");
-                }
-                other => {
-                    warn!(%context_id, delta_id = ?missing_id, ?other, "Unexpected response to delta request");
-                }
-            }
-        }
-    }
-
-    if !fetched_deltas.is_empty() {
-        info!(
-            %context_id,
-            total_fetched = fetched_deltas.len(),
-            "Adding fetched deltas to DAG in topological order"
-        );
-
-        fetched_deltas.reverse();
-
-        for (dag_delta, delta_id) in fetched_deltas {
-            if let Err(e) = delta_store.add_delta(dag_delta).await {
-                warn!(?e, %context_id, delta_id = ?delta_id, "Failed to add fetched delta to DAG");
-            }
-        }
-
-        if fetch_count > 1000 {
-            warn!(
-                %context_id,
-                total_fetched = fetch_count,
-                "Large sync detected - fetched many deltas from peer (context has deep history)"
-            );
-        }
-    }
-
-    Ok(())
-}
-
-async fn request_key_share_with_peer(
-    network_client: &NetworkClient,
-    context_client: &ContextClient,
-    context_id: &ContextId,
-    author_identity: &PublicKey,
-    peer: PeerId,
-    timeout: std::time::Duration,
-) -> Result<()> {
-    use calimero_crypto::{Nonce, SharedKey};
-    use calimero_node_primitives::sync::{InitPayload, MessagePayload, StreamMessage};
-    use rand::Rng;
-
-    debug!(
-        %context_id,
-        %author_identity,
-        %peer,
-        "Initiating bidirectional key share with peer"
-    );
-
-    tokio::time::timeout(timeout, async {
-        let mut stream = network_client.open_stream(peer).await?;
-
-        let identities = context_client.get_context_members(context_id, Some(true));
-        let Some((our_identity, _)) = choose_stream(identities, &mut rand::thread_rng())
-            .await
-            .transpose()?
-        else {
-            bail!("no owned identities found for context");
-        };
-
-        let our_nonce = rand::thread_rng().gen::<Nonce>();
-
-        crate::sync::stream::send(
-            &mut stream,
-            &StreamMessage::Init {
-                context_id: *context_id,
-                party_id: our_identity,
-                payload: InitPayload::KeyShare,
-                next_nonce: our_nonce,
-            },
-            None,
-        )
-        .await?;
-
-        let Some(ack) = crate::sync::stream::recv(&mut stream, None, timeout).await? else {
-            bail!("connection closed while awaiting key share handshake");
-        };
-
-        let their_nonce = match ack {
-            StreamMessage::Init {
-                payload: InitPayload::KeyShare,
-                next_nonce,
-                ..
-            } => next_nonce,
-            unexpected => {
-                bail!("unexpected message during key share: {:?}", unexpected)
-            }
-        };
-
-        let mut their_identity = context_client
-            .get_identity(context_id, author_identity)?
-            .ok_or_eyre("expected peer identity to exist")?;
-
-        let (private_key, sender_key) = context_client
-            .get_identity(context_id, &our_identity)?
-            .and_then(|i| Some((i.private_key?, i.sender_key?)))
-            .ok_or_eyre("expected own identity to have private & sender keys")?;
-
-        let shared_key = SharedKey::new(&private_key, &their_identity.public_key);
-
-        crate::sync::stream::send(
-            &mut stream,
-            &StreamMessage::Message {
-                sequence_id: 0,
-                payload: MessagePayload::KeyShare { sender_key },
-                next_nonce: our_nonce,
-            },
-            Some((shared_key, our_nonce)),
-        )
-        .await?;
-
-        let Some(msg) =
-            crate::sync::stream::recv(&mut stream, Some((shared_key, their_nonce)), timeout)
-                .await?
-        else {
-            bail!("connection closed while awaiting sender_key");
-        };
-
-        let their_sender_key = match msg {
-            StreamMessage::Message {
-                payload: MessagePayload::KeyShare { sender_key },
-                ..
-            } => sender_key,
-            unexpected => {
-                bail!("unexpected message: {:?}", unexpected)
-            }
-        };
-
-        their_identity.sender_key = Some(their_sender_key);
-        context_client.update_identity(context_id, &their_identity)?;
-
-        info!(
-            %context_id,
-            our_identity=%our_identity,
-            their_identity=%author_identity,
-            %peer,
-            "Bidirectional key share completed"
-        );
-
-        Ok(())
-    })
-    .await
-    .map_err(|_| eyre::eyre!("Timeout during key share with peer"))??
 }
