+++ crates/node/src/comms/direct/mod.rs
@@ -1,8 +1,8 @@
-pub mod blob_protocol;
+pub mod blob_share;
 pub mod dag_bootstrapper;
 pub mod delta_request;
 pub mod helpers;
-pub mod ingress;
+pub mod key_exchange;
 pub mod peer_selector;
 pub mod request_queue;
-pub mod stream_responder;
+pub mod streams;
