+++ crates/node/src/comms/direct/streams/mod.rs
@@ -1,3 +1,17 @@
+//! Routing and responders for direct libp2p streams.
+//!
+//! `handle_stream_opened` acts as the narrow integration point from the
+//! network layer into the sync subsystem. It simply routes incoming streams to
+//! either the blob responder (`BlobResponder`) or the sync responder
+//! (`SyncResponder`). The concrete responders live in submodules to keep this
+//! file focused on the orchestration logic.
+
+mod blob;
+mod sync;
+
+pub(crate) use blob::BlobResponder;
+pub(crate) use sync::SyncResponder;
+
 use actix::{AsyncContext, WrapFuture};
 use calimero_network_primitives::stream::Stream;
 use libp2p::{PeerId, StreamProtocol};
@@ -5,8 +19,8 @@ use tracing::{debug, info};
 
 use crate::NodeManager;
 
-use super::blob_protocol::handle_blob_protocol_stream;
-
+/// Routes a newly-opened libp2p stream to either the blob protocol or the
+/// general sync responder.
 pub fn handle_stream_opened(
     node_manager: &mut NodeManager,
     ctx: &mut <NodeManager as actix::Actor>::Context,
@@ -15,18 +29,21 @@ pub fn handle_stream_opened(
     protocol: StreamProtocol,
 ) {
     if protocol == calimero_network_primitives::stream::CALIMERO_BLOB_PROTOCOL {
-        info!(%peer_id, "Routing to blob protocol handler");
-        let node_client = node_manager.clients.node.clone();
+        info!(%peer_id, "Routing to blob protocol responder");
+        let responder = BlobResponder::new(
+            node_manager.clients.node.clone(),
+            node_manager.clients.context.clone(),
+        );
         let _ignored = ctx.spawn(
             async move {
-                if let Err(err) = handle_blob_protocol_stream(node_client, peer_id, stream).await {
+                if let Err(err) = responder.handle_stream(peer_id, stream).await {
                     debug!(%peer_id, error = %err, "Failed to handle blob protocol stream");
                 }
             }
             .into_actor(node_manager),
         );
     } else {
-        debug!(%peer_id, "Routing to sync protocol handler");
+        debug!(%peer_id, "Routing to sync protocol responder");
         let sync_manager = node_manager.managers.sync.clone();
         let _ignored = ctx.spawn(
             async move {
