+++ crates/node/src/comms/direct/streams/sync.rs
@@ -1,38 +1,36 @@
 use calimero_context_primitives::client::ContextClient;
-use calimero_crypto::Nonce;
 use calimero_network_primitives::stream::Stream;
-use calimero_node_primitives::client::NodeClient;
-use calimero_node_primitives::sync::{InitPayload, StreamMessage};
-use calimero_primitives::context::ContextId;
+use calimero_node_primitives::sync::direct::{InitPayload, StreamMessage};
 use eyre::{bail, Result};
-use tracing::{error, info};
+use rand::thread_rng;
+use tracing::error;
 
-use crate::sync::blobs;
-use crate::sync::config::SyncConfig;
-use crate::sync::key;
+use super::BlobResponder;
+use crate::comms::direct::{delta_request, key_exchange};
 use crate::sync::stream::{recv, send};
-use crate::sync::tracking::Sequencer;
+use crate::sync::SyncConfig;
+use crate::utils::choose_stream;
 
 #[derive(Clone, Debug)]
-pub(crate) struct StreamResponder {
+pub(crate) struct SyncResponder {
     sync_config: SyncConfig,
-    node_client: NodeClient,
     context_client: ContextClient,
     node_state: crate::NodeState,
+    blob_responder: BlobResponder,
 }
 
-impl StreamResponder {
+impl SyncResponder {
     pub(crate) fn new(
         sync_config: SyncConfig,
-        node_client: NodeClient,
         context_client: ContextClient,
         node_state: crate::NodeState,
+        blob_responder: BlobResponder,
     ) -> Self {
         Self {
             sync_config,
-            node_client,
             context_client,
             node_state,
+            blob_responder,
         }
     }
 
@@ -57,7 +55,7 @@ impl StreamResponder {
             return Ok(None);
         };
 
-        let (context_id, their_identity, payload, nonce) = match message {
+        let (context_id, their_identity, payload, _nonce) = match message {
             StreamMessage::Init {
                 context_id,
                 party_id,
@@ -98,113 +96,61 @@ impl StreamResponder {
             .context_client
             .get_context_members(&context.id, Some(true));
 
-        let Some((our_identity, _)) =
-            crate::utils::choose_stream(identities, &mut rand::thread_rng())
-                .await
-                .transpose()?
+        let Some((our_identity, _)) = choose_stream(identities, &mut thread_rng())
+            .await
+            .transpose()?
         else {
             bail!("no owned identities found for context: {}", context.id);
         };
 
         match payload {
             InitPayload::KeyShare => {
-                self.handle_key_share_request(&context, our_identity, their_identity, stream, nonce)
-                    .await?
-            }
-            InitPayload::BlobShare { blob_id } => {
-                self.handle_blob_share_request(
+                key_exchange::handle_key_share_request(
+                    &self.context_client,
                     &context,
                     our_identity,
                     their_identity,
-                    blob_id,
                     stream,
+                    self.sync_config.timeout,
                 )
                 .await?
+            }
+            InitPayload::BlobShare { blob_id } => {
+                self.blob_responder
+                    .handle_blob_share_request(
+                        &context,
+                        our_identity,
+                        their_identity,
+                        blob_id,
+                        stream,
+                    )
+                    .await?
             }
             InitPayload::DeltaRequest {
                 context_id: requested_context_id,
                 delta_id,
             } => {
-                self.handle_delta_request(requested_context_id, delta_id, stream)
-                    .await?
+                delta_request::handle_delta_request(
+                    &self.context_client,
+                    &self.node_state,
+                    requested_context_id,
+                    delta_id,
+                    stream,
+                )
+                .await?
             }
             InitPayload::DagHeadsRequest {
                 context_id: requested_context_id,
             } => {
-                self.handle_dag_heads_request(requested_context_id, stream)
-                    .await?
+                delta_request::handle_dag_heads_request(
+                    &self.context_client,
+                    requested_context_id,
+                    stream,
+                )
+                .await?
             }
         };
 
         Ok(Some(()))
-    }
-
-    async fn handle_key_share_request(
-        &self,
-        context: &calimero_context_primitives::types::Context,
-        our_identity: calimero_primitives::identity::PublicKey,
-        their_identity: calimero_primitives::identity::PublicKey,
-        stream: &mut Stream,
-        nonce: Nonce,
-    ) -> Result<()> {
-        key::handle_key_share_request(
-            &self.node_client,
-            &self.context_client,
-            context,
-            our_identity,
-            their_identity,
-            stream,
-            nonce,
-        )
-        .await
-    }
-
-    async fn handle_blob_share_request(
-        &self,
-        context: &calimero_context_primitives::types::Context,
-        our_identity: calimero_primitives::identity::PublicKey,
-        their_identity: calimero_primitives::identity::PublicKey,
-        blob_id: calimero_primitives::blobs::BlobId,
-        stream: &mut Stream,
-    ) -> Result<()> {
-        blobs::handle_blob_share_request(
-            &self.node_client,
-            &self.context_client,
-            context,
-            our_identity,
-            their_identity,
-            blob_id,
-            stream,
-        )
-        .await
-    }
-
-    async fn handle_delta_request(
-        &self,
-        context_id: ContextId,
-        delta_id: [u8; 32],
-        stream: &mut Stream,
-    ) -> Result<()> {
-        crate::sync::direct::delta_request::handle_delta_request(
-            &self.context_client,
-            &self.node_state,
-            context_id,
-            delta_id,
-            stream,
-        )
-        .await
-    }
-
-    async fn handle_dag_heads_request(
-        &self,
-        context_id: ContextId,
-        stream: &mut Stream,
-    ) -> Result<()> {
-        crate::sync::direct::delta_request::handle_dag_heads_request(
-            &self.context_client,
-            context_id,
-            stream,
-        )
-        .await
     }
 }
